% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanjeevscbe.R
\name{normalize}
\alias{normalize}
\title{Performs normalization for a distribution vector}
\usage{
normalize(dis_vec)
}
\arguments{
\item{dis_vec}{A distribution vector.}
}
\value{
norm_dis_vec A distribution vector.
}
\description{
Performs normalization for a distribution vector
}
\examples{
norm_dist <- normalize(c(1,3,2,4,5,3))
norm_dist <- normalize(c(1,2,3))
}
