% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanjeevscbe.R
\name{standardize}
\alias{standardize}
\title{Performs standardization for a distribution vector}
\usage{
standardize(dis_vec)
}
\arguments{
\item{dis_vec}{A distribution vector.}
}
\value{
stan_dis_vec A distribution vector.
}
\description{
Performs standardization for a distribution vector
}
\examples{
stand_dist <- standardize(c(1,3,2,4,5,3))
stand_dist <- standardize(c(1,2,3))
}
